+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\files\assertWritableDirectory
 - Adds oihana\files\deleteDirectory
 - Adds oihana\files\deleteTemporaryDirectory
+- Adds oihana\files\directoryPath
 - Adds oihana\files\findFiles
 - Adds oihana\files\getBaseFileName
 - Adds oihana\files\getDirectory